@@ -2,11 +2,15 @@ import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { HomeComponent } from './pages/home/home.component';
 import { NotFoundComponent } from './pages/not-found/not-found.component';
-
+import { CountryDetailComponent } from './pages/country-detail/country-detail.component';
 const routes: Routes = [
   {
     path: '',
     component: HomeComponent,
+  },
+  { 
+    path: 'country-detail/:countryName', 
+    component: CountryDetailComponent 
   },
   {
     path: '**', // wildcard
