@@ -1,5 +1,5 @@
-import { Component, OnInit } from '@angular/core';
-import { take } from 'rxjs';
+import { Component, OnInit, OnDestroy } from '@angular/core';
+import { takeUntil, Subject } from 'rxjs';
 import { OlympicService } from './core/services/olympic.service';
 
 @Component({
@@ -7,10 +7,15 @@ import { OlympicService } from './core/services/olympic.service';
   templateUrl: './app.component.html',
   styleUrls: ['./app.component.scss'],
 })
-export class AppComponent implements OnInit {
+export class AppComponent implements OnInit, OnDestroy {
   constructor(private olympicService: OlympicService) {}
+  private destroy$ = new Subject<void>();
 
   ngOnInit(): void {
-    this.olympicService.loadInitialData().pipe(take(1)).subscribe();
+    this.olympicService.loadInitialData().pipe(takeUntil(this.destroy$)).subscribe();
+  }
+  ngOnDestroy(): void {
+    this.destroy$.next();
+    this.destroy$.complete()
   }
 }
