@@ -1,8 +1,8 @@
 import { Participation } from './Participation';
 
-interface Olympic {
-    id: number;
-    country: string;
-    participations: Participation[];
-  }
+export interface Olympic {
+  id: number;
+  country: string;
+  participations: Participation[];
+}
   
